Refactor: Extract app initialization and logic into modules
Extract complex initialization and component logic from App.tsx and index.tsx into separate modules and custom hooks for improved maintainability and code organization.

Changes include:
- Split App.tsx into AppRoutes and AppOverlays components
- Extract feature-specific hooks: useGenericPopup, useGlobalBrowserEffects, useGrowthBookFeatureCache, useHotUpdate, useNativeAppListeners, useOpsConsoleBodyClass, useRemoteAssetFlags, useRouteAudioCleanup, useUsageLimitModal
- Extract startup logic: errorReporting, growthbookClient, nativeRuntime, platformSetup, renderRoot, serviceBootstrap
- Simplify index.tsx to orchestrate initialization and rendering
- No functional changes, purely organizational refactoring
