+++ src/App.tsx
@@ -15,13 +15,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-import { Route, Switch, useLocation } from 'react-router-dom';
-import {
-  IonApp,
-  IonRouterOutlet,
-  IonToast,
-  setupIonicReact,
-} from '@ionic/react';
+import { IonApp, IonRouterOutlet, setupIonicReact } from '@ionic/react';
 import { IonReactRouter } from '@ionic/react-router';
 
 /* Core CSS required for Ionic components to work properly */
@@ -39,946 +33,75 @@ import '@ionic/react/css/text-alignment.css';
 import '@ionic/react/css/text-transformation.css';
 import '@ionic/react/css/flex-utils.css';
 import '@ionic/react/css/display.css';
+
 /* Theme variables */
 import './theme/variables.css';
-import Home from './pages/Home';
-import { End } from './pages/End';
-import { useEffect, useLayoutEffect, useRef, useState } from 'react';
-import { Capacitor, registerPlugin } from '@capacitor/core';
-import { Filesystem, Directory } from '@capacitor/filesystem';
-import Profile from './pages/Profile';
-import ProtectedRoute from './ProtectedRoute';
-import { App as CapApp } from '@capacitor/app';
-import {
-  // APP_LANG,
-  BASE_NAME,
-  BUNDLE_ZIP_URLS,
-  CACHE_IMAGE,
-  HOMEWORK_REMOTE_ASSETS_ENABLED,
-  CAN_ACCESS_REMOTE_ASSETS,
-  DOWNLOADING_CHAPTER_ID,
-  DOWNLOAD_BUTTON_LOADING_STATUS,
-  HOMEWORK_PATHWAY_ASSETS,
-  IS_CUBA,
-  LEARNING_PATH_ASSETS,
-  MODES,
-  PAGES,
-  PortPlugin,
-  SHOULD_SHOW_HOMEWORK_REMOTE_ASSETS,
-  SHOULD_SHOW_REMOTE_ASSETS,
-  SHOW_GENERIC_POPUP,
-  GENERIC_POP_UP,
-  SEARCH_LESSON_CACHE_KEY,
-  SEARCH_LESSON_HISTORY,
-  PAL_LEARNING_RATES_CONFIG,
-  LIDO_BUNDLE_ZIP_URLS,
-} from './common/constants';
-import { Util } from './utility/util';
-import Parent from './pages/Parent';
-import DisplayStudents from './pages/DisplayStudents';
-// import Assignments from "./pages/Assignments";
-// import { Keyboard, KeyboardResize } from "@capacitor/keyboard";
-import DisplaySubjects from './pages/DisplaySubjects';
-import AddCourses from './pages/AddCourses';
-import StudentProgress from './pages/StudentProgress';
-import SearchLesson from './pages/SearchLesson';
-import Leaderboard from './pages/Leaderboard';
-import SelectMode from './pages/SelectMode';
-import HotUpdate from './pages/HotUpdate';
-import TermsAndConditions from './pages/TermsAndConditions';
-import DisplayChapters from './pages/DisplayChapters';
-import LiveQuizRoom from './pages/LiveQuizRoom';
-import LiveQuizGame from './pages/LiveQuizGame';
-import LiveQuizRoomResult from './pages/LiveQuizRoomResult';
-import LiveQuizLeaderBoard from './pages/LiveQuizLeaderBoard';
-import { useOnlineOfflineErrorMessageHandler } from './common/onlineOfflineErrorMessageHandler';
-import { t } from 'i18next';
-// import TeacherProfile from "./pages/Malta/TeacherProfile";
-import React from 'react';
 import './App.css';
-import { schoolUtil } from './utility/schoolUtil';
-import LidoPlayer from './pages/LidoPlayer';
-import Loading from './components/Loading';
-import UploadPage from './ops-console/pages/UploadPage';
-import SidebarPage from './ops-console/pages/SidebarPage';
-import { initializeClickListener } from './analytics/clickUtil';
-import ResetPassword from './pages/ResetPassword';
-import DisplayClasses from './teachers-module/pages/DisplayClasses';
-import LessonDetails from './teachers-module/pages/LessonDetails';
-import ShowStudentsInAssignmentPage from './teachers-module/pages/ShowStudentsInAssignmentPage';
-import ReqEditSchool from './teachers-module/pages/ReqEditSchool';
-import StudentProfile from './teachers-module/pages/StudentProfile';
-import AddStudent from './teachers-module/pages/AddStudent';
-import UserProfile from './teachers-module/pages/UserProfile';
-import SubjectSelection from './teachers-module/pages/SubjectSelection';
-import DisplaySchools from './teachers-module/pages/DisplaySchools';
-import StudentReport from './teachers-module/pages/StudentReport';
-import ManageSchools from './teachers-module/pages/ManageSchools';
-import SchoolProfile from './teachers-module/pages/SchoolProfile';
-import ManageClass from './teachers-module/pages/ManageClass';
-import DashBoardDetails from './teachers-module/pages/DashBoardDetails';
-import EditClass from './teachers-module/pages/EditClass';
-import ClassProfile from './teachers-module/pages/ClassProfile';
-import ShowChapters from './teachers-module/pages/ShowChapters';
-import SearchLessons from './teachers-module/pages/SearchLessons';
-import HomePage from './teachers-module/pages/HomePage';
-import TeacherLibraryAssignments from './teachers-module/pages/TeacherLibraryAssignments';
-import ClassUsers from './teachers-module/pages/ClassUsers';
-import AddTeacher from './teachers-module/pages/AddTeacher';
-import TeacherProfile from './teachers-module/pages/TeacherProfile';
-import SchoolUsers from './teachers-module/pages/SchoolUsers';
-import AddSchoolUser from './teachers-module/pages/AddSchoolUser';
-import ProgramsPage from './ops-console/pages/ProgramPage';
-import SchoolList from './ops-console/pages/SchoolList';
-import { useFeatureValue, useFeatureIsOn } from '@growthbook/growthbook-react';
-import LoginScreen from './pages/LoginScreen';
-import ProfileDetails from './components/profileDetails/ProfileDetails';
-import RequestList from './ops-console/pages/RequestList';
-import AddTeacherName from './teachers-module/pages/AddTeacherName';
-import SearchSchool from './teachers-module/pages/SearchSchool';
-import JoinSchool from './pages/JoinSchool';
-import CreateSchool from './teachers-module/pages/CreateSchool';
-import ScanRedirect from './teachers-module/components/homePage/assignment/ScanRedirect';
-import GenericPopup from './components/GenericPopUp/GenericPopUp';
-import PopupManager from './components/GenericPopUp/GenericPopUpManager';
-import TermsGate from './components/termsandconditons/TermsGate';
-import { useGrowthBook } from '@growthbook/growthbook-react';
-import { setCachedGrowthBookFeatureValue } from './growthbook/Growthbook';
-import { useAppSelector } from './redux/hooks';
-import { HardwareBackButtonHandler } from './common/backButtonRegistry';
-import { logger } from './utility/logger';
-import {
-  getBundleZipUrlsForEnv,
-  getLidoBundleZipUrlsForEnv,
-} from './services/RemoteConfig';
-import {
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogActions,
-  Button,
-} from '@mui/material';
 
-import ColoringBoard from './components/coloring/ColoringBoard';
-import PostSuccess from './teachers-module/pages/PostSuccess';
-import QRAssignments from './teachers-module/components/homePage/assignment/QRAssignments';
-import TeacherRecommendedAssignments from './teachers-module/components/homePage/assignment/TeacherRecommendedAssignments';
-import StreakPage from './teachers-module/components/streakComponent/streakPage';
-import StickerBook from './pages/StickerBook';
-import KidsAppLocation from './teachers-module/pages/KidsAppLocation';
-import { AudioUtil } from './utility/AudioUtil';
+import React from 'react';
+import { BASE_NAME } from './common/constants';
+import TermsGate from './components/termsandconditons/TermsGate';
+import { HardwareBackButtonHandler } from './common/backButtonRegistry';
+import { useAppSelector } from './redux/hooks';
 import { useNavigationHandler } from './helper/navigation/NavigationHandler';
+import AppRoutes from './app/AppRoutes';
+import AppOverlays from './app/AppOverlays';
+import { useGenericPopup } from './hooks/useGenericPopup';
+import { useGlobalBrowserEffects } from './hooks/useGlobalBrowserEffects';
+import { useGrowthBookFeatureCache } from './hooks/useGrowthBookFeatureCache';
+import { useHotUpdate } from './hooks/useHotUpdate';
+import { useNativeAppListeners } from './hooks/useNativeAppListeners';
+import { useOpsConsoleBodyClass } from './hooks/useOpsConsoleBodyClass';
+import { useRemoteAssetFlags } from './hooks/useRemoteAssetFlags';
+import { useRouteAudioCleanup } from './hooks/useRouteAudioCleanup';
+import { useUsageLimitModal } from './hooks/useUsageLimitModal';
 
 setupIonicReact();
-interface ExtraData {
-  notificationType?: string;
-  rewardProfileId?: string;
-  classId?: string;
-}
-interface WindowEventMap {
-  shouldShowModal: CustomEvent<boolean>;
-}
-type GrowthBookJsonConfig = Record<string, unknown>;
-type GrowthBookFeatureDebugResult<T> = {
-  value: T | null;
-  source: string;
-};
 
-const TIME_LIMIT = 1500; // 25 * 60
-const LAST_MODAL_SHOWN_KEY = 'lastTimeExceededShown';
-const START_TIME_KEY = 'startTime';
-const USED_TIME_KEY = 'usedTime';
-const LAST_ACCESS_DATE_KEY = 'lastAccessDate';
-const IS_INITIALIZED = 'isInitialized';
-let timeoutId: NodeJS.Timeout;
-
-const RouteAudioCleanup = () => {
-  const location = useLocation();
-  const lastLocationRef = useRef(location.pathname + location.search);
-
-  useEffect(() => {
-    const currentLocation = location.pathname + location.search;
-    if (lastLocationRef.current !== currentLocation) {
-      void AudioUtil.stopAudioUrlOrTtsPlayback();
-      lastLocationRef.current = currentLocation;
-    }
-  }, [location.pathname, location.search]);
-
-  return null;
-};
-
-const NavigationHandler = () => {
+const AppRouteEffects = () => {
   useNavigationHandler();
+  useOpsConsoleBodyClass();
+  useRouteAudioCleanup();
   return null;
 };
 
 const App: React.FC = () => {
-  const growthbook = useGrowthBook();
-  const [online, setOnline] = useState(navigator.onLine);
-  const { presentToast } = useOnlineOfflineErrorMessageHandler();
-  const [startTime, setStartTime] = useState<number>(() => {
-    const savedStartTime = localStorage.getItem(START_TIME_KEY);
-    const initialTime = savedStartTime ? Number(savedStartTime) : Date.now();
-    if (!savedStartTime) {
-      localStorage.setItem('startTime', initialTime.toString());
-    }
-    return initialTime;
-  });
-  const [popupData, setPopupData] = useState<any>(null);
-  const [timeExceeded, setTimeExceeded] = useState<boolean>(false);
-  const [showModal, setShowModal] = useState<boolean>(false);
-  const [showToast, setShowToast] = useState<boolean>(false);
   const isGlobalLoading = useAppSelector((state) => state.auth.globalLoading);
-  const [isActive, setIsActive] = useState(true);
-  const shouldShowRemoteAssets = useFeatureIsOn(CAN_ACCESS_REMOTE_ASSETS);
-  const shouldShowHomeworkRemoteAssets = useFeatureIsOn(
-    HOMEWORK_REMOTE_ASSETS_ENABLED,
-  );
+  const popup = useGenericPopup();
+  const usageLimit = useUsageLimitModal();
 
-  const popupDataRef = useRef<any>(null);
-  const showModalRef = useRef(showModal);
-  const bundleZipUrlsFallbackRef = useRef(getBundleZipUrlsForEnv());
-  const lidoBundleZipUrlsFallbackRef = useRef(getLidoBundleZipUrlsForEnv());
+  useGrowthBookFeatureCache();
+  useRemoteAssetFlags();
+  useGlobalBrowserEffects();
+  useNativeAppListeners();
+  useHotUpdate();
 
-  useEffect(() => {
-    popupDataRef.current = popupData;
-  }, [popupData]);
-
-  useEffect(() => {
-    showModalRef.current = showModal;
-  }, [showModal]);
-
-  const learningPathAssets: any = useFeatureValue(LEARNING_PATH_ASSETS, {});
-  const homeworkPathwayAssets: any = useFeatureValue(
-    HOMEWORK_PATHWAY_ASSETS,
-    {},
-  );
-  const palLearningRatesConfig = useFeatureValue<GrowthBookJsonConfig>(
-    PAL_LEARNING_RATES_CONFIG,
-    {},
-  );
-  const bundleZipUrls = useFeatureValue<string[]>(
-    BUNDLE_ZIP_URLS,
-    bundleZipUrlsFallbackRef.current,
-  );
-  const lidoBundleZipUrls = useFeatureValue<string[]>(
-    LIDO_BUNDLE_ZIP_URLS,
-    lidoBundleZipUrlsFallbackRef.current,
-  );
-
-  const OpsConsoleRouteWatcher = () => {
-    const location = useLocation();
-
-    useEffect(() => {
-      const isOpsConsoleRoute = location.pathname.includes(PAGES.SIDEBAR_PAGE);
-
-      if (isOpsConsoleRoute) {
-        document.body.classList.add('ops-console');
-      } else {
-        document.body.classList.remove('ops-console');
-      }
-
-      return () => {
-        document.body.classList.remove('ops-console');
-      };
-    }, [location.pathname]);
-
-    return null;
-  };
-
-  useEffect(() => {
-    // this event listener is to remove the highlighted text(if exists) on a click
-    const handleClick = () => {
-      const sel = window.getSelection();
-      if (sel && !sel.isCollapsed) {
-        sel.removeAllRanges();
-      }
-    };
-    document.addEventListener('click', handleClick);
-    localStorage.removeItem(SEARCH_LESSON_CACHE_KEY);
-    localStorage.removeItem(SEARCH_LESSON_HISTORY);
-    return () => {
-      document.removeEventListener('click', handleClick);
-    };
-  }, []);
-
-  useEffect(() => {
-    if (
-      palLearningRatesConfig &&
-      typeof palLearningRatesConfig === 'object' &&
-      Object.keys(palLearningRatesConfig).length > 0
-    ) {
-      setCachedGrowthBookFeatureValue(
-        PAL_LEARNING_RATES_CONFIG,
-        palLearningRatesConfig,
-      );
-    }
-  }, [palLearningRatesConfig]);
-
-  useEffect(() => {
-    if (!growthbook) return;
-    const getFeatureDebugResult = <T,>(
-      featureKey: string,
-      resolvedValue: T,
-      fallbackValue: T,
-    ): GrowthBookFeatureDebugResult<T> => {
-      if (typeof growthbook?.evalFeature === 'function') {
-        const result = growthbook.evalFeature<T>(featureKey);
-        return {
-          value: result?.value ?? null,
-          source: result?.source ?? 'unknown',
-        };
-      }
-
-      return {
-        value: resolvedValue,
-        source: resolvedValue === fallbackValue ? 'fallback-or-mock' : 'mocked',
-      };
-    };
-
-    const bundleZipUrlsResult = getFeatureDebugResult(
-      BUNDLE_ZIP_URLS,
-      bundleZipUrls,
-      bundleZipUrlsFallbackRef.current,
-    );
-    const lidoBundleZipUrlsResult = getFeatureDebugResult(
-      LIDO_BUNDLE_ZIP_URLS,
-      lidoBundleZipUrls,
-      lidoBundleZipUrlsFallbackRef.current,
-    );
-
-    logger.warn('[GrowthBook] bundle ZIP URLs evaluated', {
-      featureKey: BUNDLE_ZIP_URLS,
-      growthBookSource: bundleZipUrlsResult.source,
-      growthBookValue: bundleZipUrlsResult.value,
-      fallbackValue: bundleZipUrlsFallbackRef.current,
-      resolvedValue: bundleZipUrls,
-      usingGrowthBookValue: bundleZipUrlsResult.value !== null,
-    });
-
-    logger.warn('[GrowthBook] Lido bundle ZIP URLs evaluated', {
-      featureKey: LIDO_BUNDLE_ZIP_URLS,
-      growthBookSource: lidoBundleZipUrlsResult.source,
-      growthBookValue: lidoBundleZipUrlsResult.value,
-      fallbackValue: lidoBundleZipUrlsFallbackRef.current,
-      resolvedValue: lidoBundleZipUrls,
-      usingGrowthBookValue: lidoBundleZipUrlsResult.value !== null,
-    });
-
-    setCachedGrowthBookFeatureValue(BUNDLE_ZIP_URLS, bundleZipUrls);
-    setCachedGrowthBookFeatureValue(LIDO_BUNDLE_ZIP_URLS, lidoBundleZipUrls);
-  }, [growthbook, bundleZipUrls, lidoBundleZipUrls]);
-
-  useEffect(() => {
-    if (!growthbook) return;
-
-    const popupConfig = growthbook.getFeatureValue(GENERIC_POP_UP, null) as any;
-
-    if (!popupConfig) return;
-
-    const params = new URLSearchParams(window.location.search);
-    const currentTab = params.get('tab');
-
-    if (
-      currentTab &&
-      popupConfig.screen_name &&
-      currentTab.toLowerCase() === popupConfig.screen_name.toLowerCase()
-    ) {
-      PopupManager.onAppOpen(popupConfig);
-      PopupManager.onTimeElapsed(popupConfig);
-    }
-  }, [growthbook, window.location.search]);
-
-  useLayoutEffect(() => {
-    const handler = (e: any) => {
-      logger.info('POPUP EVENT RECEIVED', e.detail);
-      setPopupData(e.detail);
-    };
-
-    window.addEventListener(SHOW_GENERIC_POPUP, handler);
-
-    return () => {
-      window.removeEventListener(SHOW_GENERIC_POPUP, handler);
-    };
-  }, []);
-
-  useEffect(() => {
-    const handler = (e: any) => {
-      logger.info('POPUP EVENT', e.detail);
-      setPopupData(e.detail);
-    };
-
-    window.addEventListener(SHOW_GENERIC_POPUP, handler);
-    return () => window.removeEventListener(SHOW_GENERIC_POPUP, handler);
-  }, []);
-
-  useEffect(() => {
-    const cleanup = initializeClickListener();
-    const handleOnline = () => {
-      if (!online) {
-        setOnline(true);
-        presentToast({
-          message: 'Device is online.',
-          color: 'success',
-          duration: 3000,
-          position: 'bottom',
-          buttons: [
-            {
-              text: 'Dismiss',
-              role: 'cancel',
-            },
-          ],
-        });
-      }
-    };
-
-    const handleOffline = () => {
-      setOnline(false);
-      presentToast({
-        message: 'Device is offline.',
-        color: 'danger',
-        duration: 3000,
-        position: 'bottom',
-        buttons: [
-          {
-            text: 'Dismiss',
-            role: 'cancel',
-          },
-        ],
-      });
-    };
-
-    window.addEventListener('online', handleOnline);
-    window.addEventListener('offline', handleOffline);
-
-    return () => {
-      window.removeEventListener('online', handleOnline);
-      window.removeEventListener('offline', handleOffline);
-      cleanup();
-    };
-  }, [online, presentToast]);
-  useEffect(() => {
-    initializeUsage();
-    document.addEventListener('visibilitychange', handleVisibilityChange);
-    startTimeout();
-    localStorage.setItem(DOWNLOAD_BUTTON_LOADING_STATUS, JSON.stringify(false));
-    localStorage.setItem(DOWNLOADING_CHAPTER_ID, JSON.stringify(false));
-    CapApp.addListener('appStateChange', Util.onAppStateChange);
-    localStorage.setItem(IS_CUBA, '1');
-    if (Capacitor.isNativePlatform()) {
-      //CapApp.addListener("appStateChange", Util.onAppStateChange);
-      // Keyboard.setResizeMode({ mode: KeyboardResize.Ionic });
-
-      const portPlugin = registerPlugin<PortPlugin>('Port');
-      portPlugin.addListener('notificationOpened', (data: any) => {
-        if (data) {
-          processNotificationData(data);
-        }
-      });
-      CapApp.addListener('appUrlOpen', Util.onAppUrlOpen);
-    }
-
-    if (shouldShowRemoteAssets) {
-      Util.DownloadRemoteAssets(
-        learningPathAssets?.asset_repo_url,
-        learningPathAssets?.uniqueId,
-        'remoteAsset', // The destination folder
-        'Learning Path', // The asset type for logging
-      );
-    }
-    localStorage.setItem(
-      SHOULD_SHOW_REMOTE_ASSETS,
-      JSON.stringify(shouldShowRemoteAssets),
-    );
-
-    if (shouldShowHomeworkRemoteAssets) {
-      Util.DownloadRemoteAssets(
-        homeworkPathwayAssets?.asset_repo_url,
-        homeworkPathwayAssets?.uniqueId,
-        'homeworkRemoteAsset', // The DIFFERENT destination folder
-        'Homework', // The asset type for logging
-      );
-    }
-    localStorage.setItem(
-      SHOULD_SHOW_HOMEWORK_REMOTE_ASSETS,
-      JSON.stringify(shouldShowHomeworkRemoteAssets),
-    );
-
-    if (Capacitor.isNativePlatform()) {
-      try {
-        Filesystem.mkdir({
-          path: CACHE_IMAGE,
-          directory: Directory.Cache,
-        }).catch((e) => {
-          throw new Error('Error in creating directory for cache');
-        });
-      } catch (e) {
-        logger.error('Error in creating directory for cache', e);
-      }
-    }
-    //Checking for flexible update in play-store
-    Util.startFlexibleUpdate();
-
-    //Listen to network change
-    Util.listenToNetwork();
-    fetchData();
-
-    Util.notificationListener(async (extraData: ExtraData | undefined) => {
-      if (extraData) {
-        Util.navigateTabByNotificationData(extraData);
-      }
-    });
-    updateAvatarSuggestionJson();
-    // Cleanup on unmount
-    return () => {
-      document.removeEventListener('visibilitychange', handleVisibilityChange);
-      clearExistingTimeout();
-    };
-  }, []);
-
-  const initializeUsage = () => {
-    const currentDate = new Date().toISOString().split('T')[0];
-    const lastAccessDate = localStorage.getItem(LAST_ACCESS_DATE_KEY);
-
-    if (!lastAccessDate || lastAccessDate !== currentDate) {
-      // First-time use or a new day
-      localStorage.setItem(USED_TIME_KEY, '0'); // Reset used time
-      localStorage.setItem(START_TIME_KEY, Date.now().toString()); // Reset start time
-      localStorage.setItem(LAST_ACCESS_DATE_KEY, currentDate); // Update the last access date
-    }
-
-    if (!localStorage.getItem(IS_INITIALIZED)) {
-      localStorage.setItem(START_TIME_KEY, Date.now().toString());
-      localStorage.setItem(IS_INITIALIZED, 'true');
-    }
-  };
-
-  // Function to calculate the used time and store it
-  const calculateUsedTime = () => {
-    const currentTime = Date.now();
-    const startTime = Number(
-      localStorage.getItem(START_TIME_KEY) || currentTime,
-    ); // Use current time if startTime is missing
-    const usedTime = Number(localStorage.getItem(USED_TIME_KEY));
-    const sessionTime = (currentTime - startTime) / 1000;
-    const usedTimeInMinutes = usedTime / 60;
-
-    return usedTime + sessionTime;
-  };
-
-  const saveUsedTime = () => {
-    const totalUsedTime = calculateUsedTime();
-    localStorage.setItem(USED_TIME_KEY, totalUsedTime.toString());
-  };
-
-  const startTimeout = () => {
-    clearExistingTimeout();
-    const usedTime = Number(localStorage.getItem(USED_TIME_KEY) || 0);
-    const remainingTime = Util.TIME_LIMIT - usedTime;
-    if (remainingTime > 0) {
-      timeoutId = setTimeout(() => {
-        checkTimeExceeded();
-      }, remainingTime * 1000);
-    }
-  };
-
-  const clearExistingTimeout = () => {
-    clearTimeout(timeoutId);
-  };
-  const checkTimeExceeded = async () => {
-    if (Capacitor.isNativePlatform()) {
-      const currMode = await schoolUtil.getCurrMode();
-      if (currMode === MODES.PARENT) {
-        const today = new Date().toISOString().split('T')[0];
-        const lastModalShownDate = localStorage.getItem(LAST_MODAL_SHOWN_KEY);
-
-        if (lastModalShownDate !== today) {
-          setShowModal(true);
-          const event = new CustomEvent('shouldShowModal', { detail: true });
-          window.dispatchEvent(event);
-          localStorage.setItem(LAST_MODAL_SHOWN_KEY, today);
-        }
-      }
-    }
-  };
-
-  // Function to handle visibility change (when app goes into background or foreground)
-  const handleVisibilityChange = () => {
-    const currentTime = Date.now();
-    if (document.visibilityState === 'visible') {
-      if (!localStorage.getItem(START_TIME_KEY)) {
-        localStorage.setItem(START_TIME_KEY, currentTime.toString());
-      }
-      startTimeout();
-    } else {
-      void AudioUtil.stopAudioUrlOrTtsPlayback();
-      saveUsedTime();
-      localStorage.removeItem(START_TIME_KEY);
-      clearExistingTimeout();
-    }
-  };
-
-  const handleContinue = () => {
-    setShowModal(false);
-    setShowToast(true);
-    setStartTime(Date.now());
-    localStorage.setItem(START_TIME_KEY, Date.now().toString());
-  };
-  const processNotificationData = async (data: ExtraData | undefined) => {
-    Util.navigateTabByNotificationData(data);
-  };
-  const getNotificationData = async () => {
-    if (Capacitor.isNativePlatform()) {
-      if (!Util.port) Util.port = registerPlugin<PortPlugin>('Port');
-      if (Util.port && typeof Util.port.fetchNotificationData === 'function') {
-        try {
-          const data = await Util.port.fetchNotificationData();
-          if (data) {
-            processNotificationData(data);
-          }
-        } catch (error) {
-          logger.error('Error retrieving notification data:', error);
-        }
-      } else {
-        logger.warn('Util.port or fetchNotificationData is not available.');
-      }
-    }
-  };
-  const fetchData = async () => {
-    await getNotificationData();
-  };
-
-  async function updateAvatarSuggestionJson() {
-    // Update Avatar Suggestion local Json
-    try {
-    } catch (error) {
-      logger.error('Util.migrateLocalJsonFile failed ', error);
-    }
-  }
   return (
     <IonApp>
       <IonReactRouter basename={BASE_NAME}>
-        <NavigationHandler />
-        <OpsConsoleRouteWatcher />
-        <RouteAudioCleanup />
+        <AppRouteEffects />
         <TermsGate />
         <HardwareBackButtonHandler
-          popupDataRef={popupDataRef}
-          setPopupData={setPopupData}
-          popupManager={PopupManager}
-          showModalRef={showModalRef}
-          setShowModal={setShowModal}
+          popupDataRef={popup.popupDataRef}
+          setPopupData={popup.setPopupData}
+          popupManager={popup.popupManager}
+          showModalRef={usageLimit.showModalRef}
+          setShowModal={usageLimit.setShowModal}
         />
         <IonRouterOutlet>
-          <Switch>
-            <Route path={PAGES.APP_UPDATE} exact={true}>
-              <HotUpdate />
-            </Route>
-            <Route path={PAGES.RESET_PASSWORD} exact={true}>
-              <ResetPassword />
-            </Route>
-            <ProtectedRoute path={PAGES.HOME} exact={true}>
-              <Home />
-            </ProtectedRoute>
-            <Route path={PAGES.LOGIN} exact={true}>
-              <LoginScreen />
-            </Route>
-            <ProtectedRoute path={PAGES.LIDO_PLAYER} exact={true}>
-              <LidoPlayer />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.END} exact={true}>
-              <End />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.PROFILE} exact={true}>
-              <Profile />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.PARENT} exact={true}>
-              <Parent />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.QR_ASSIGNMENTS} exact={true}>
-              <QRAssignments />
-            </ProtectedRoute>
-            {/* <Route path={PAGES.APP_LANG_SELECTION} exact={true}>
-              <AppLangSelection />
-            </Route> */}
-            <ProtectedRoute path={PAGES.CREATE_STUDENT} exact={true}>
-              <ProfileDetails />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.EDIT_STUDENT} exact={true}>
-              <ProfileDetails />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.DISPLAY_STUDENT} exact={true}>
-              <DisplayStudents />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.DISPLAY_SUBJECTS} exact={true}>
-              <DisplaySubjects />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.ADD_SUBJECTS} exact={true}>
-              <AddCourses />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.DISPLAY_CHAPTERS} exact={true}>
-              <DisplayChapters />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.STUDENT_PROGRESS} exact={true}>
-              <StudentProgress />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.SEARCH} exact={true}>
-              <SearchLesson />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.LEADERBOARD} exact={true}>
-              <Leaderboard />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.ASSIGNMENT} exact={true}>
-              <Home />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.LIVE_QUIZ} exact={true}>
-              <Home />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.JOIN_CLASS} exact={true}>
-              <Home />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.JOIN_SCHOOL} exact={true}>
-              <JoinSchool />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.CREATE_SCHOOL} exact={true}>
-              <CreateSchool />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.SELECT_MODE} exact={true}>
-              <SelectMode />
-            </ProtectedRoute>
-            {/* <ProtectedRoute path={PAGES.TEACHER_PROFILE} exact={true}>
-              <TeacherProfile />
-            </ProtectedRoute> */}
-            <ProtectedRoute path={PAGES.STUDENT_PROFILE} exact={true}>
-              <StudentProfile />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.ADD_STUDENT} exact={true}>
-              <AddStudent />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.USER_PROFILE} exact={true}>
-              <UserProfile />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.ADD_TEACHER_NAME} exact={true}>
-              <AddTeacherName />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.SUBJECTS_PAGE} exact={true}>
-              <SubjectSelection />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.LIVE_QUIZ_JOIN} exact={true}>
-              <LiveQuizRoom />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.LIVE_QUIZ_GAME} exact={true}>
-              <LiveQuizGame />
-            </ProtectedRoute>
-            <Route path={PAGES.TERMS_AND_CONDITIONS} exact={true}>
-              <TermsAndConditions />
-            </Route>
-            <ProtectedRoute path={PAGES.LIVE_QUIZ_ROOM_RESULT} exact={true}>
-              <LiveQuizRoomResult />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.LIVE_QUIZ_LEADERBOARD} exact={true}>
-              <LiveQuizLeaderBoard />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.DISPLAY_SCHOOLS} exact={true}>
-              <DisplaySchools />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.KIDS_APP_LOCATION} exact={true}>
-              <KidsAppLocation />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.SEARCH_SCHOOL} exact={true}>
-              <SearchSchool />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.STUDENT_REPORT} exact={true}>
-              <StudentReport />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.DISPLAY_CLASSES} exact={true}>
-              <DisplayClasses />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.MANAGE_SCHOOL} exact={true}>
-              <ManageSchools />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.SCHOOL_PROFILE} exact={true}>
-              <SchoolProfile />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.COLORING_BOARD} exact={true}>
-              <ColoringBoard />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.STICKER_BOOK} exact={true}>
-              <StickerBook />
-            </ProtectedRoute>
-            {/* <ProtectedRoute path={PAGES.ADD_SCHOOL} exact={true}>
-
-                <EditSchool />
-
-            </ProtectedRoute> */}
-            <ProtectedRoute path={PAGES.REQ_ADD_SCHOOL} exact={true}>
-              <ReqEditSchool />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.SCAN_REDIRECT}>
-              <ScanRedirect />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.MANAGE_CLASS} exact={true}>
-              <ManageClass />
-            </ProtectedRoute>
-            {/* <ProtectedRoute path={PAGES.EDIT_SCHOOL} exact={true}>
-
-                <EditSchool />
-
-            </ProtectedRoute> */}
-            <ProtectedRoute path={PAGES.REQ_EDIT_SCHOOL} exact={true}>
-              <ReqEditSchool />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.DASHBOARD_DETAILS} exact={true}>
-              <DashBoardDetails />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.ADD_CLASS} exact={true}>
-              <EditClass />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.CLASS_PROFILE} exact={true}>
-              <ClassProfile />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.SHOW_CHAPTERS} exact={true}>
-              <ShowChapters />
-            </ProtectedRoute>
-
-            <ProtectedRoute path={PAGES.SEARCH_LESSON} exact={true}>
-              <SearchLessons />
-            </ProtectedRoute>
-
-            <ProtectedRoute path={PAGES.LESSON_DETAILS} exact={true}>
-              <LessonDetails />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.HOME_PAGE} exact={true}>
-              <HomePage />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.STREAK_PAGE} exact={true}>
-              <StreakPage />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.TEACHER_ASSIGNMENT} exact={true}>
-              <TeacherLibraryAssignments />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.POST_SUCCESS} exact={true}>
-              <PostSuccess />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.CLASS_USERS} exact={true}>
-              <ClassUsers />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.EDIT_CLASS} exact={true}>
-              <EditClass />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.SCHOOL_LIST} exact={true}>
-              <SchoolList />
-            </ProtectedRoute>
-            <ProtectedRoute
-              path={PAGES.TEACHER_RECOMMENDED_ASSIGNMENTS}
-              exact={true}
-            >
-              <TeacherRecommendedAssignments />
-            </ProtectedRoute>
-            <ProtectedRoute
-              path={PAGES.SHOW_STUDENTS_IN_ASSIGNED_PAGE}
-              exact={true}
-            >
-              <ShowStudentsInAssignmentPage />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.ADD_TEACHER} exact={true}>
-              <AddTeacher />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.TEACHER_PROFILE} exact={true}>
-              <TeacherProfile />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.SCHOOL_USERS} exact={true}>
-              <SchoolUsers />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.ADD_PRINCIPAL} exact={true}>
-              <AddSchoolUser />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.ADD_COORDINATOR} exact={true}>
-              <AddSchoolUser />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.ADD_SPONSOR} exact={true}>
-              <AddSchoolUser />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.UPLOAD_PAGE} exact={true}>
-              <UploadPage />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.PROGRAM_PAGE} exact={true}>
-              <ProgramsPage />
-            </ProtectedRoute>
-            <ProtectedRoute path={PAGES.REQUEST_LIST} exact={true}>
-              <RequestList />
-            </ProtectedRoute>
-
-            <ProtectedRoute path={PAGES.SIDEBAR_PAGE}>
-              <SidebarPage />
-            </ProtectedRoute>
-          </Switch>
+          <AppRoutes />
         </IonRouterOutlet>
-
-        <Dialog
-          open={showModal}
-          onClose={(event, reason) => {
-            if (reason === 'backdropClick' || reason === 'escapeKeyDown') {
-              // prevent closing
-              return;
-            }
-            handleContinue();
-          }}
-          className="custom-dialog"
-        >
-          <DialogTitle sx={{ textAlign: 'center' }}>
-            {t('Time for a break!') || ''}
-          </DialogTitle>
-          <DialogContent sx={{ textAlign: 'center' }}>
-            {t(
-              'You’ve used Chimple for 25 minutes today. Take a break to rest your eyes!',
-            ) || ''}
-          </DialogContent>
-          <DialogActions sx={{ justifyContent: 'center' }}>
-            <Button
-              variant="contained"
-              color="success"
-              onClick={handleContinue}
-              sx={{
-                borderRadius: '1vh',
-                padding: '1vh 2vw',
-                minWidth: '20vh',
-                fontWeight: 'bold',
-              }}
-            >
-              {t('Continue')}
-            </Button>
-          </DialogActions>
-        </Dialog>
-
-        {/*Toast notification for acknowledgment */}
-        <IonToast
-          isOpen={showToast}
-          onDidDismiss={() => setShowToast(false)}
-          message="You have resumed after exceeding the time limit."
-          duration={3000}
+        <AppOverlays
+          isGlobalLoading={isGlobalLoading}
+          popupData={popup.popupData}
+          onPopupClose={popup.closePopup}
+          onPopupAction={popup.actOnPopup}
+          showBreakModal={usageLimit.showModal}
+          onContinueFromBreak={usageLimit.continueAfterBreak}
+          showBreakToast={usageLimit.showToast}
+          onDismissBreakToast={() => usageLimit.setShowToast(false)}
         />
       </IonReactRouter>
-      {popupData && (
-        <GenericPopup
-          thumbnailImageUrl={popupData.localized.thumbnailImageUrl}
-          backgroundImageUrl={popupData.localized.backgroundImageUrl}
-          audioUrl={popupData.localized.audioUrl}
-          heading={popupData.localized.heading}
-          subHeading={popupData.localized.subHeading}
-          details={popupData.localized.details}
-          buttonText={popupData.localized.buttonText}
-          onClose={() => {
-            PopupManager.onDismiss(popupData.config);
-            setPopupData(null);
-          }}
-          onAction={() => {
-            PopupManager.onAction(popupData.config);
-            setPopupData(null);
-          }}
-        />
-      )}
-      <Loading isLoading={isGlobalLoading} />
     </IonApp>
   );
 };
