+++ src/startup/platformSetup.ts
@@ -1,0 +1,51 @@
+import { defineCustomElements as jeepSqlite } from 'jeep-sqlite/loader';
+import { defineCustomElements, JSX as LocalJSX } from 'lido-standalone/loader';
+import {
+  SpeechSynthesis,
+  SpeechSynthesisUtterance,
+} from '../utility/WindowsSpeech';
+import { initializeFireBase } from '../services/Firebase';
+import { isNativePlatform } from './nativeRuntime';
+
+type NavigatorWithUserAgentData = Navigator & {
+  userAgentData?: {
+    mobile?: boolean;
+  };
+};
+
+declare global {
+  namespace JSX {
+    interface IntrinsicElements extends LocalJSX.IntrinsicElements {}
+  }
+}
+
+export const initializePlatformSetup = () => {
+  applyMobileWebBrowserClass();
+  defineCustomElements(window);
+  initializeFireBase();
+  initializeSpeechPolyfills();
+  jeepSqlite(window);
+};
+
+const applyMobileWebBrowserClass = () => {
+  const userAgent = navigator.userAgent || '';
+  const userAgentData = (navigator as NavigatorWithUserAgentData).userAgentData;
+  const isMobileBrowser =
+    userAgentData?.mobile === true || /\bMobile\b/i.test(userAgent);
+
+  document.body.classList.toggle(
+    'mobile-web-browser',
+    !isNativePlatform && isMobileBrowser,
+  );
+};
+
+const initializeSpeechPolyfills = () => {
+  if (typeof window === 'undefined') return;
+
+  if (!(window as any).speechSynthesis) {
+    (window as any).speechSynthesis = new SpeechSynthesis();
+  }
+  if (!(window as any).SpeechSynthesisUtterance) {
+    (window as any).SpeechSynthesisUtterance = SpeechSynthesisUtterance;
+  }
+};
